Orange: DEV-3766: Incorporate PR feedback
